requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
langchain>=0.3.0
langchain-core==0.3.29
langchain-ollama==0.2.2
langchain-groq>=0.2.0
python-dotenv==1.0.1
requests>=2.31.0
rdflib==7.1.4
